Implementing DepartmentSectionViewModel
