+++ TheMET/TheMET/ViewControllers/DepartmentsSectionViewController.swift
@@ -8,10 +8,16 @@
 import Foundation
 import UIKit
 import MetAPI
+import Combine
 
 class DepartmentsSectionViewController: UIViewController {
     
     private let metAPI = MetAPI()
+    
+    private var viewModel: DepartmentsSectionViewModel?
+    
+    private var contentStatusSubscriber: AnyCancellable?
+    private var titleSubscriber: AnyCancellable?
     
     private let departmentLabel: UILabel = UILabel()
     
@@ -19,73 +25,51 @@ class DepartmentsSectionViewController: UIViewController {
     private let failedCatalogView = CatalogSectionTapToReloadView.constractView(configuration: .catalogSectionTapToReload)
     private let loadedCatalogView = DepartmentsSectionContentView.constructView()
     
-    private var loadedDepartments: [Department] = []
-    
-    var contentStatus:LoadingStatus<[CatalogSectionCellData<Int>]> = .loading {
-        didSet {
-            self.updateContent()
-        }
-    }
-    
-    var loadingDepartmentIds: [Int] = []
-    
     override func viewDidLoad() {
         super.viewDidLoad()
         self.departmentLabel.translatesAutoresizingMaskIntoConstraints = false
         self.view.addSubview(departmentLabel)
-        self.departmentLabel.apply(font: NSLocalizedString("serif_font", comment: ""), color: UIColor(named: "pear"), fontSize: 16, title: NSLocalizedString("catalog_screen_departments_section_title", comment: ""))
         self.departmentLabel.textAlignment = .center
         NSLayoutConstraint.activate([
             self.departmentLabel.leadingAnchor.constraint(equalTo: self.view.safeAreaLayoutGuide.leadingAnchor, constant: 20),
             self.departmentLabel.trailingAnchor.constraint(equalTo: self.view.safeAreaLayoutGuide.trailingAnchor, constant: -20),
             self.departmentLabel.topAnchor.constraint(equalTo: self.view.topAnchor),
             self.departmentLabel.heightAnchor.constraint(equalToConstant: 60)
-            ])
+        ])
         self.add(catalogSubview: self.failedCatalogView)
         self.add(catalogSubview: self.loadedCatalogView)
         self.add(catalogSubview: self.loadingCatalogView)
         self.failedCatalogView.onButtonTap = { [weak self] in
-            self?.reloadButtonDidTap()
+            self?.viewModel?.reloadButtonDidTap()
         }
         self.loadedCatalogView.onCatalogCellTap = { [weak self] departmentId in
-            self?.catalogCellDidTap(departmentId)
+            self?.viewModel?.cellDidTap(departmentId: departmentId)
         }
         self.loadedCatalogView.onCatalogCellWillDisplay = { [weak self] departmentId in
-            self?.loadParticularDepartmentCellDataIfNeeded(departmentId: departmentId)
+            self?.viewModel?.loadParticularDepartmentCellDataIfNeeded(departmentId: departmentId)
         }
-        self.reloadCatalog()
+        self.setupViewModel()
+        self.viewModel?.reloadCatalog()
     }
     
-    private func loadParticularDepartmentCellDataIfNeeded(departmentId: Int) {
-        guard case .loaded(let catalogCellDataList) = self.contentStatus else {
-            return
-        }
-        guard let catalogCellData = catalogCellDataList.first(where: { cellData in
-            cellData.identificator == departmentId
-        }),
-              case .placeholder = catalogCellData.data,
-              !self.loadingDepartmentIds.contains(departmentId),
-        let department = self.loadedDepartments.first(where: { department in
-            department.id == departmentId
+    private func setupViewModel() {
+        let viewModel = DepartmentsSectionViewModel(presentingControllerProvider: { [weak self] in
+            return self
         })
-        else {
-            return
-        }
-        self.loadingDepartmentIds.append(departmentId)
-        self.loadCatalogCellData(department: department, completion: { [weak self] catalogCellData in
-               if let contentStatus = self?.contentStatus,
-               case .loaded(var catalogCellDataList) = contentStatus,
-                  let catalogCellDataIndex = catalogCellDataList.firstIndex(where: { $0.identificator == catalogCellData.identificator}) {
-                catalogCellDataList[catalogCellDataIndex] = catalogCellData
-                self?.contentStatus = .loaded(catalogCellDataList)
-            }
-            self?.loadingDepartmentIds.removeAll { id in
-                return id == departmentId
-            }
-        })
+        self.viewModel = viewModel
+        self.titleSubscriber = viewModel.$title
+            .sink(receiveValue: { [weak self] titleText in
+                guard let self = self else {return}
+                self.departmentLabel.apply(font: NSLocalizedString("serif_font", comment: ""), color: UIColor(named: "pear"), fontSize: 16, title: titleText)
+            })
+        self.contentStatusSubscriber = viewModel.$contentStatus
+            .sink(receiveValue: { [weak self] contentStatus in
+                guard let self = self else {return}
+                self.updateContent(contentStatus: contentStatus)
+            })
     }
     
-    private func updateContent() {
+    private func updateContent(contentStatus: LoadingStatus<[CatalogSectionCellData<Int>]>) {
         switch contentStatus {
         case .failed:
             self.loadedCatalogView.content = []
@@ -117,89 +101,15 @@ class DepartmentsSectionViewController: UIViewController {
     }
     
     private func reloadCatalog() {
-        self.contentStatus = .loading
-        self.metAPI.departments { [weak self] departmentResponseResult in
-            switch departmentResponseResult {
-            case .failure(let error):
-                self?.contentStatus = .failed(error)
-            case .success(let departmentResponse):
-                var catalogCellDataList: [CatalogSectionCellData<Int>] = []
-                for department in departmentResponse.departments {
-                    let catalogCellData: CatalogSectionCellData = CatalogSectionCellData(identificator: department.id, data: .placeholder)
-                    catalogCellDataList.append(catalogCellData)
-                }
-                self?.loadedDepartments = departmentResponse.departments
-                self?.contentStatus = .loaded(catalogCellDataList)
-            }
-        }
-    }
-    
-    private func loadCatalogCellData(department: Department, completion: @escaping (CatalogSectionCellData<Int>) -> Void) {
-        self.metAPI.objects(departmentIds: [department.id]) { [weak self] objectsResult in
-            guard let self = self else {
-                return
-            }
-            switch objectsResult {
-            case .failure:
-                let catalogCellData = CatalogSectionCellData(identificator: department.id, data: .data(imageURL: nil, title: department.displayName, subTitle: nil))
-                completion(catalogCellData)
-            case .success(let objects):
-                let subtitle: String = self.cellSubtitle(objectsCount: objects.total)
-                self.loadDepartmentImageURL(objectIds: objects.objectIDs, completion: { url in
-                    let catalogCellData = CatalogSectionCellData(identificator: department.id, data: .data(imageURL: url, title: department.displayName, subTitle: subtitle))
-                    completion(catalogCellData)
-                })
-            }
-        }
-    }
-    
-    
-    
-    private func cellSubtitle(objectsCount: ArtID) -> String {
-        let formatString: String = NSLocalizedString("objects count", comment: "")
-        return String.localizedStringWithFormat(formatString, objectsCount)
-    }
-    
-    private func loadDepartmentImageURL(objectIds: [ArtID], completion: @escaping (URL?) -> Void) {
-        guard !objectIds.isEmpty else {
-            completion(nil)
-            return
-        }
-        self.loadObjectImageURL(objectId: objectIds[0]) { [weak self] url in
-            if let imageURL = url {
-                completion(imageURL)
-            } else {
-                let newObjectsIds: [ArtID] = Array(objectIds.dropFirst())
-                self?.loadDepartmentImageURL(objectIds: newObjectsIds, completion: completion)
-            }
-        }
-    }
-    
-    private func loadObjectImageURL(objectId: ArtID, completion: @escaping (URL?) -> Void){
-        self.metAPI.object(id: objectId) { objectResult in
-            switch objectResult {
-            case .failure:
-                completion(nil)
-            case .success(let object):
-                guard let imageURLstring = object.primaryImage else {
-                    completion(nil)
-                    return
-                }
-                completion(URL(string: imageURLstring))
-            }
-        }
-    }
-    
-    private func reloadButtonDidTap() {
-        self.reloadCatalog()
+        let viewModel = DepartmentsSectionViewModel(presentingControllerProvider: { [weak self] in
+            return self
+        })
+        self.viewModel = viewModel
+        viewModel.reloadCatalog()
     }
     
     private func catalogCellDidTap(_ departmentId: Int) {
-        let mainStoryBoard = UIStoryboard(name: "Main", bundle: nil)
-        let categoryViewController = mainStoryBoard.instantiateViewController(withIdentifier: "CategoryViewController") as? CategoryViewController
-        guard let categoryViewController = categoryViewController else { return }
-        categoryViewController.departmentId = departmentId
-        self.navigationController?.pushViewController(categoryViewController, animated: true)
+        self.viewModel?.cellDidTap(departmentId: departmentId)
     }
     
 }
